+++ src/iptux/TransFileModel.cpp
@@ -4,6 +4,14 @@
 #include <glib.h>
 
 namespace iptux {
+
+TransFileModel::TransFileModel()
+: fileLength(0),
+  finishedLength(0),
+  finished(false),
+  data(nullptr) {
+}
+
 TransFileModel& TransFileModel::setStatus(const std::string& value) {
   status = value;
   return *this;
@@ -60,8 +68,14 @@ TransFileModel& TransFileModel::setFilePath(const std::string& value) {
 }
 
 TransFileModel& TransFileModel::setData(TransAbstract* value) {
+  g_assert_nonnull(value);
   data = value;
   return *this;
+}
+
+void TransFileModel::finish() {
+  finished = true;
+  data = nullptr;
 }
 
 const std::string& TransFileModel::getStatus() const {
@@ -131,6 +145,10 @@ TransAbstract* TransFileModel::getData() const {
 }
 int64_t TransFileModel::getFileLength() const {
   return fileLength;
+}
+
+bool TransFileModel::isFinished() const {
+  return finished;
 }
 
 }