+++ src/iptux/UiUtils.cpp
@@ -1,5 +1,7 @@
 #include "UiUtils.h"
 
+#include "iptux/UiModels.h"
+
 namespace iptux {
 
 void UiUtils::applyTransFileModel2GtkListStore(
@@ -8,20 +10,21 @@ void UiUtils::applyTransFileModel2GtkListStore(
     GtkTreeIter *iter) {
   gtk_list_store_set(
       list_store, iter,
-      0, para.getStatus().c_str(),
-      1, para.getTask().c_str(),
-      2, para.getPeer().c_str(),
-      3, para.getIp().c_str(),
-      4, para.getFilename().c_str(),
-      5, para.getFileLengthText().c_str(),
-      6, para.getFinishedLengthText().c_str(),
-      7, int(para.getProgress()),
-      8, g_strdup(para.getProgressText().c_str()),
-      9, para.getCost().c_str(),
-      10, para.getRemain().c_str(),
-      11, para.getRate().c_str(),
-      12, para.getFilePath().c_str(),
-      13, para.getData(),
+      TransModelColumn::STATUS, para.getStatus().c_str(),
+      TransModelColumn::TASK, para.getTask().c_str(),
+      TransModelColumn::PEER, para.getPeer().c_str(),
+      TransModelColumn::IP, para.getIp().c_str(),
+      TransModelColumn::FILENAME, para.getFilename().c_str(),
+      TransModelColumn::FILE_LENGTH_TEXT, para.getFileLengthText().c_str(),
+      TransModelColumn::FINISHED_LENGTH_TEXT, para.getFinishedLengthText().c_str(),
+      TransModelColumn::PROGRESS, int(para.getProgress()),
+      TransModelColumn::PROGRESS_TEXT, g_strdup(para.getProgressText().c_str()),
+      TransModelColumn::COST, para.getCost().c_str(),
+      TransModelColumn::REMAIN, para.getRemain().c_str(),
+      TransModelColumn::RATE, para.getRate().c_str(),
+      TransModelColumn::FILE_PATH, para.getFilePath().c_str(),
+      TransModelColumn::DATA, para.getData(),
+      TransModelColumn::FINISHED, para.isFinished(),
       -1);
 }
 }